refactor(references): замечания ревью и Sonar по фиксу переиндексации
- fill() снимает содержимое документа ДО чтения AST и пишет отпечаток из
  снимка: конкурентное перестроение больше не может оставить отпечаток
  новее проиндексированного AST (залипание устаревшего индекса исключено,
  лишняя переиндексация безопасна) — по замечанию ревью;
- политика доступности ссылки вынесена в ReferenceAccessibility —
  снижение связанности ReferenceIndex (java:S1200) и отдельная зона
  ответственности;
- стресс-тест считает исключения потоков-читателей и проваливается при
  их наличии — по замечанию ревью.
